% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_uk_cols.R
\name{separate_uk_cols}
\alias{separate_uk_cols}
\title{separate_uk_cols}
\usage{
separate_uk_cols(data, col, pattern = "[^[:alnum:]]+", into_prefix)
}
\arguments{
\item{data}{data.frame or tibble}

\item{col}{column to be separated}

\item{pattern}{regular expression pattern}

\item{into_prefix}{prefix for new columns}
}
\description{
separate_uk_cols
}
